superpower/powersource/requirements.txt
click
cherrypy
flask
itsdangerous>=1.1.0
Flask-Session>=0.3.1
flask-talisman
Flask-Caching>=1.4.0
Flask-Opentracing==0.2.0
jaeger-client==3.11.0
opentracing<2,>=1.2.2
prometheus-client>=0.4.2
prometheus-flask-exporter>=0.3.1
wsgi-request-logger
simplejson
python-slugify
uwsgi